requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-appconf==1.0.3
django-countries==6.0
django-ipware==2.1.0
django-rest-framework==0.1.0
django-tracking-analyzer==1.0.1
django-user-agents==0.4.0
djangorestframework==3.11.0
geoip2==3.0.0
idna==2.9
maxminddb==1.5.2
numpy==1.18.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
ua-parser==0.10.0
urllib3==1.25.8
user-agents==2.1
